% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul_by_two.R
\name{mul_by_two}
\alias{mul_by_two}
\title{Multiple a number by two}
\usage{
mul_by_two(num)
}
\arguments{
\item{num}{A number}
}
\value{
Multiple of a number by 2
}
\description{
Multiple a number by two
}
\examples{
mul_by_two(3)
}
